+++ upcloud/testdata/upcloud_managed_database/managed_database_s1.tf
@@ -164,6 +164,15 @@ resource "upcloud_managed_database_user" "db_user_2" {
   }
 }
 
+resource "upcloud_managed_database_user" "db_user_3" {
+  service  = upcloud_managed_database_postgresql.pg2.id
+  username = "autopassword"
+
+  pg_access_control {
+    allow_replication = true
+  }
+}
+
 resource "upcloud_managed_database_user" "db_user_4" {
   service  = upcloud_managed_database_opensearch.o1.id
   username = "somename"