+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [5.36.0] - 2026-03-20
+
 ### Added
 
 - upcloud_loadbalancer: Add http_rewrite_path and http_rewrite_uri actions
@@ -1190,7 +1192,8 @@ Updated upcloud-go-api, added build/CI scripts, and repackaged 0.1.0 as 1.0.0.
 - resource_upcloud_firewall_rule removed and replaced by resource_upcloud_firewall_rules
 - resource_upcloud_zone removed and replaced by zone and zones datasources
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.35.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.36.0...HEAD
+[5.36.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.35.0...v5.36.0
 [5.35.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.34.1...v5.35.0
 [5.34.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.34.0...v5.34.1
 [5.34.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.33.4...v5.34.0